chore(typescript): consolidated the .d.ts and .js files
