+++ packages/victory-core/src/victory-primitives/clip-path.tsx
@@ -1,0 +1,27 @@
+import React from "react";
+import PropTypes from "prop-types";
+import { VictoryCommonPrimitiveProps } from "../victory-util/types";
+
+export interface ClipPathProps extends VictoryCommonPrimitiveProps {
+  children?: React.ReactNode[] | React.ReactNode;
+  clipId?: number | string;
+}
+
+const ClipPath = (props: ClipPathProps) => (
+  <defs>
+    {
+      // @ts-expect-error FIXME: "id cannot be a number"
+      <clipPath id={props.clipId}>{props.children}</clipPath>
+    }
+  </defs>
+);
+
+ClipPath.propTypes = {
+  children: PropTypes.oneOfType([
+    PropTypes.arrayOf(PropTypes.node),
+    PropTypes.node
+  ]),
+  clipId: PropTypes.oneOfType([PropTypes.number, PropTypes.string])
+};
+
+export default ClipPath;
