--- packages/victory-core/src/victory-primitives/line.d.ts
@@ -1,4 +1,0 @@
-import * as React from "react";
-import { VictoryPrimitiveShapeProps } from "./types";
-
-export default class Line extends React.Component<VictoryPrimitiveShapeProps> {}
