--- packages/victory-core/src/victory-primitives/rect.d.ts
@@ -1,4 +1,0 @@
-import * as React from "react";
-import { VictoryPrimitiveShapeProps } from "./types";
-
-export default class Rect extends React.Component<VictoryPrimitiveShapeProps> {}
