--- packages/victory-core/src/victory-primitives/text.d.ts
@@ -1,10 +1,0 @@
-import { VictoryCommonPrimitiveProps } from "../victory-util/types";
-import * as React from "react";
-
-export interface TextProps extends VictoryCommonPrimitiveProps {
-  children?: React.ReactNode;
-  desc?: string;
-  title?: string;
-}
-
-export default class Text extends React.Component<TextProps> {}
