--- packages/victory-core/src/victory-primitives/tspan.d.ts
@@ -1,4 +1,0 @@
-import * as React from "react";
-import { VictoryCommonPrimitiveProps } from "../victory-util/types";
-
-export default class TSpan extends React.Component<VictoryCommonPrimitiveProps> {}
