--- packages/victory-core/src/victory-primitives/tspan.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-const TSpan = (props) => <tspan {...props} />;
-
-export default TSpan;
